pyproject.toml
[tool.poetry]
name = "py-ptc"
version = "0.1.0"
description = ""
authors = ["Dnevb <villalbautria08@gmail.com>"]

[tool.poetry.scripts]
py-ptc = "py_ptc.cli:main"

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.0.1"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
